feat: add blogs MongoDB connection and functionality
