+++ bootcamp-milestone-2/.env
@@ -1,1 +1,1 @@
-MONGO_URI = "mongodb+srv://blammatthew:zIybu8bHg7vwhegr@cluster0.2mbko.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0"
+MONGO_URI = "mongodb+srv://blammatthew:zIybu8bHg7vwhegr@cluster0.2mbko.mongodb.net/blogsDB?retryWrites=true&w=majority&appName=Cluster0"
