+++ bootcamp-milestone-2/src/app/blogData.ts
@@ -12,7 +12,7 @@ const blogs: Blog[] = [
         title: "The Calc Free-Rider",
         date: "10/16/24",
         description: "It was supposed to be a PARTNER quiz >:|",
-        image: ".../public/images/calc.jpg",
+        image: "/images/calc.jpg",
         imageAlt: "Calculus problems",
         slug: "calcfreerider",
     },
@@ -20,7 +20,7 @@ const blogs: Blog[] = [
         title: "Hack4Impact Meeting",
         date: "10/17/24",
         description: "Hack4Impact bootcamp meeting",
-        image: ".../public/images/meeting.jpg",
+        image: "/images/meeting.jpg",
         imageAlt: "React presentation",
         slug: "hack4impact",
     },
